+++ plugins/expo/src/expo.ts
@@ -9,6 +9,7 @@ import {
 } from "@hot-updater/plugin-core";
 import { ExecaError, execa } from "execa";
 import { resolveMain } from "./resolveMain";
+import { runExpoPrebuild } from "./util/prebuild";
 
 interface RunBundleArgs {
   cwd: string;
@@ -130,6 +131,11 @@ export const expo =
   ({ cwd }: BasePluginArgs): BuildPlugin => {
     const { outDir = "dist", sourcemap = false } = config;
     return {
+      nativeBuild: {
+        prebuild: async ({ platform }) => {
+          await runExpoPrebuild({ platform });
+        },
+      },
       build: async ({ platform }) => {
         const buildPath = path.join(cwd, outDir);