+++ packages/core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @hot-updater/core
 
+## 0.19.7
+
+## 0.19.6
+
+## 0.19.5
+
+### Patch Changes
+
+- 40d28c2: bump rnef
+
 ## 0.19.4
 
 ## 0.19.3