+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @hot-updater/react-native
 
+## 0.19.7
+
+### Patch Changes
+
+- Updated dependencies [e28313d]
+- Updated dependencies [bcc641e]
+  - hot-updater@0.19.7
+  - @hot-updater/core@0.19.7
+  - @hot-updater/js@0.19.7
+
+## 0.19.6
+
+### Patch Changes
+
+- 657a10e: Android Native Build - Gradle Build
+- Updated dependencies [657a10e]
+  - hot-updater@0.19.6
+  - @hot-updater/core@0.19.6
+  - @hot-updater/js@0.19.6
+
+## 0.19.5
+
+### Patch Changes
+
+- 40d28c2: bump rnef
+- d3ac760: fix: delete previous bundle when previous bundle access is needed
+- Updated dependencies [40d28c2]
+  - @hot-updater/core@0.19.5
+  - hot-updater@0.19.5
+  - @hot-updater/js@0.19.5
+
 ## 0.19.4
 
 ### Patch Changes